add: change add nft flow
