+++ components/collection/HeroButtonAddNfts.vue
@@ -10,11 +10,13 @@ import { NeoDropdownItem } from '@kodadot1/brick'
 const route = useRoute()
 
 const id = route.params.id.toString()
+const { urlPrefix } = usePrefix()
 
 const addNfts = () => {
   navigateTo({
-    path: '/create/nft',
+    path: `/${urlPrefix.value}/create`,
     query: {
+      select: 'nft',
       collectionId: id,
     },
   })
